# bruno/openapi.json
{"openapi":"3.1.0","info":{"title":"Flux","summary":"API endpoints for the Flux Central hub","description":"\n# Sections\n## Integrations\nThe first set of tags are around interacting with the integrations.\n\n## Operational\nThis set of tags is split by actions. Get, Upsert, or Delete\n\n## Core\nThis set is for the core functionality of the app.\n","version":"0.1.0"},"paths":{"/general/health":{"get":{"tags":["general","get"],"summary":"Get App Health","description":"Endpoint for health check\nReturns:\n    dict: status:okay but could be anything","operationId":"get_app_health_general_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Get App Health General Health Get"}}}}}}},"/data/all":{"get":{"tags":["data","get"],"summary":"Get Database Data","description":"Endpoint for fetching database data\nReturns:\n    dict: database contents","operationId":"get_database_data_data_all_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Get Database Data Data All Get"}}}}}}},"/data/upsert":{"post":{"tags":["data","upsert"],"summary":"Upsert Database Entry","operationId":"upsert_database_entry_data_upsert_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpsertRequest"}}},"required":true},"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/MappingDoc"},{"$ref":"#/components/schemas/EncryptedCredential"}],"title":"Response Upsert Database Entry Data Upsert Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/data/{group}/{name}":{"get":{"tags":["data","get"],"summary":"Get Database Entry","operationId":"get_database_entry_data__group___name__get","parameters":[{"name":"group","in":"path","required":true,"schema":{"type":"string","title":"Group"}},{"name":"name","in":"path","required":true,"schema":{"type":"string","title":"Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MappingDoc"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/data/{doc_id}":{"delete":{"tags":["data","delete"],"summary":"Delete Database Entry","operationId":"delete_database_entry_data__doc_id__delete","parameters":[{"name":"doc_id","in":"path","required":true,"schema":{"type":"string","title":"Doc Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MappingDoc"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/traggo/version":{"get":{"tags":["traggo"],"summary":"Get Traggo Version","operationId":"get_traggo_version_traggo_version_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Get Traggo Version Traggo Version Get"}}}}}}},"/":{"get":{"tags":["general"],"summary":"Get Root","description":"Root Endpoint","operationId":"get_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Get Root  Get"}}}}}}}},"components":{"schemas":{"EncryptedCredential":{"properties":{"_id":{"anyOf":[{"$ref":"#/components/schemas/PydanticObjectId"},{"type":"null"}],"description":"MongoDB document ObjectID"},"service":{"type":"string","title":"Service"},"encrypted_api_key":{"type":"string","contentMediaType":"application/octet-stream","title":"Encrypted Api Key"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["service","encrypted_api_key"],"title":"EncryptedCredential"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Integrations":{"properties":{"traggo":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Traggo"},"sp":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sp"},"anytype":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Anytype"}},"type":"object","title":"Integrations"},"MappingDoc":{"properties":{"_id":{"anyOf":[{"$ref":"#/components/schemas/PydanticObjectId"},{"type":"null"}],"description":"MongoDB document ObjectID"},"name":{"type":"string","title":"Name"},"group":{"type":"string","title":"Group"},"integrations":{"$ref":"#/components/schemas/Integrations"}},"type":"object","required":["name","group","integrations"],"title":"MappingDoc","description":"Basic shape for Mongo document\n\nAttributes:\n    name (str): pretty name\n    type (str): project or tag\n    intrgration (dict): list of integration:id mappings"},"NewDoc":{"properties":{"name":{"type":"string","title":"Name"},"group":{"type":"string","title":"Group"},"int_name":{"type":"string","title":"Int Name"},"int_id":{"type":"string","title":"Int Id"}},"type":"object","required":["name","group","int_name","int_id"],"title":"NewDoc"},"NewKey":{"properties":{"service":{"type":"string","title":"Service"},"key":{"type":"string","title":"Key"}},"type":"object","required":["service","key"],"title":"NewKey"},"PydanticObjectId":{"type":"string","maxLength":24,"minLength":24,"pattern":"^[0-9a-f]{24}$","example":"5eb7cf5a86d9755df3a6c593"},"UpsertRequest":{"properties":{"incoming":{"anyOf":[{"$ref":"#/components/schemas/NewDoc"},{"$ref":"#/components/schemas/NewKey"}],"title":"Incoming"}},"type":"object","required":["incoming"],"title":"UpsertRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}},"tags":[{"name":"traggo","description":"Endpoints for Traggo logging","externalDocs":{"description":"Traggo docs","url":"https://traggo.net/terminology/"}},{"name":"super-productivity","description":"Endpoints for Super Productivity automation"},{"name":"anytype","description":"Endpoints for anytype interaction","externalDocs":{"description":"Anytype API docs","url":"https://developers.anytype.io/docs/reference"}},{"name":"get","description":"Endpoints for fetching objects"},{"name":"upsert","description":"Endpoints for creating or updating objects"},{"name":"delete","description":"Endpoints for deleting objects"},{"name":"data","description":"Endpoints for interacting with the mongo database"},{"name":"general","description":"Endpoints for general use"},{"name":"scheduled","description":"Which endpoints are also scheduled jobs"}]}